Refactored integration tests to improve checks & created a helper files for common functions
